Unite non-gui and GUI
-Both non-gui and GUI are now in the same program
-Menu screen at the start user can select if they want gui or non-gui
-Both games still work
