+++ classes/inventory.py
