+++ classes/magic.py
