--- rpg-gui/classes/game.py
@@ -1,49 +1,0 @@
-import random
-from .magic import Spell
-
-
-class Person:
-    def __init__(self, hp, mp, atk, df, magic, items):
-        self.maxhp = hp
-        self.hp = hp
-        self.maxmp = mp
-        self.mp = mp
-        self.atkh = atk + 10
-        self.atkl = atk - 10
-        self.df = df
-        self.magic = magic
-        self.items = items
-        self.action = ["Attack", "Magic", "Items"]
-
-    def generate_damage(self):
-        return random.randrange(self.atkl, self.atkh)
-
-    def take_damage(self, dmg):
-        self.hp -= dmg
-        if self.hp < 0:
-            self.hp = 0
-        return self.hp
-
-    def heal(self, dmg):
-        self.hp += dmg
-        if self.hp > self.maxhp:
-            self.hp = self.maxhp
-        return self.hp
-
-    def get_hp(self):
-        return self.hp
-
-    def get_maxhp(self):
-        return self.maxhp
-
-    def get_mp(self):
-        return self.mp
-
-    def get_maxmp(self):
-        return self.maxmp
-
-    def reduce_mp(self, cost):
-        self.mp -= cost
-
-    def get_action_name(self, i):
-        return self.action[i]
