--- rpg-gui/classes/inventory.py
@@ -1,6 +1,0 @@
-class Item:
-    def __init__(self, name, type, description, prop):
-        self.name = name
-        self.type = type
-        self.description = description
-        self.prop = prop
